Separated logging in with Kano World account
Among copy chages, logging in with Kano World was moved to another window making it
more clear as to what was about to happen. Now, when clicking on 'Add a new user' a
window with 2 buttons pops up from which you choose whether to login with Kano World
account or create a brand new account. The former option brings up a new view with an
entry for username and one for password. The entries now use a placeholder text instead
of labels separately.
