+++ kano_greeter/greeter_window.py
@@ -20,6 +20,7 @@ from kano.gtk3.buttons import OrangeButton
 from kano_greeter.user_list import UserList
 from kano_greeter.password_view import PasswordView
 from kano_greeter.newuser_view import NewUserView
+from kano_greeter.login_with_kw_view import LoginWithKanoWorldView
 
 
 class GreeterWindow(ApplicationWindow):
@@ -31,8 +32,8 @@ class GreeterWindow(ApplicationWindow):
     def __init__(self):
         apply_common_to_screen()
 
-        self.a=self.b=self.c=0
-        self.switching=0
+        self.a = self.b = self.c = 0
+        self.switching = 0
 
         ApplicationWindow.__init__(self, _('Login'), self.WIDTH, self.HEIGHT)
         self.connect("delete-event", Gtk.main_quit)
@@ -43,6 +44,7 @@ class GreeterWindow(ApplicationWindow):
         # Create the two views: one for normal Login, the other to create a new account
         self.password_view = PasswordView(None, self.greeter)
         self.newuser_view = NewUserView(self.greeter)
+        self.login_with_kw_view = LoginWithKanoWorldView(self.greeter)
 
         self.grid = Gtk.Grid()
         self.set_main_widget(self.grid)
@@ -83,7 +85,6 @@ class GreeterWindow(ApplicationWindow):
         self.top_bar.disable_prev()
 
     def go_to_password(self, user):
-        # Called when we switch between views using top-left arrow button
         self.set_main(self.password_view)
         self.top_bar.enable_prev()
 
@@ -93,14 +94,13 @@ class GreeterWindow(ApplicationWindow):
             self.greeter.disconnect(self.c)
 
             self.greeter = GreeterWindow.greeter.new()
-            self.password_view.greeter=self.greeter
-            (self.a,self.b,self.c) = self.password_view._reset_greeter()
-            self.switching=1
-        
+            self.password_view.greeter = self.greeter
+            (self.a, self.b, self.c) = self.password_view._reset_greeter()
+            self.switching = 1
+
         self.password_view.grab_focus(user)
 
     def go_to_newuser(self):
-        # Called when we switch between views using top-left arrow button
         self.set_main(self.newuser_view)
         self.top_bar.enable_prev()
 
@@ -110,15 +110,33 @@ class GreeterWindow(ApplicationWindow):
             self.greeter.disconnect(self.c)
 
             self.greeter = GreeterWindow.greeter.new()
-            self.newuser_view.greeter=self.greeter
+            self.newuser_view.greeter = self.greeter
 
             # FIXME: We do not reset the greeter in the newuser view
             # It will be done only when the local user account has been forced
             # so as to avoid LightDM freezes
-            #(self.a,self.b,self.c) = self.newuser_view._reset_greeter()
-            self.switching=2
+            # (self.a, self.b, self.c) = self.newuser_view._reset_greeter()
+            self.switching = 2
 
-        self.newuser_view.grab_focus()
+    def go_to_login_with_kw(self):
+        self.set_main(self.login_with_kw_view)
+        self.top_bar.enable_prev()
+
+        if not self.switching == 3:
+            self.greeter.disconnect(self.a)
+            self.greeter.disconnect(self.b)
+            self.greeter.disconnect(self.c)
+
+            self.greeter = GreeterWindow.greeter.new()
+            self.login_with_kw_view.greeter = self.greeter
+
+            # FIXME: We do not reset the greeter in the newuser view
+            # It will be done only when the local user account has been forced
+            # so as to avoid LightDM freezes
+            (self.a, self.b, self.c) = self.login_with_kw_view._reset_greeter()
+            self.switching = 3
+
+        self.login_with_kw_view.grab_focus()
 
     def _back_cb(self, event, button):
         self.go_to_users()
